+++ phase0/Scenario Walkthrough.md
@@ -5,11 +5,11 @@ The user runs the program and is prompted with the following messages:
 2. Enter Name (Alphanumeric)
 3. Enter Address (Alphanumeric)
 4. Enter a Short Summary (Optional) (String)
-5. Create a New Request. Enter "Yes" or "No"
+5. Create a New Entities.Request. Enter "Yes" or "No"
 6. Enter Product Name (Alphanumeric)
 7. Enter Quantity in Kilograms (Numeric)
 8. Enter Price per Kilogram (Numeric)
-9. Choose the number of the Distributor you wish to sell to.
+9. Choose the number of the Entities.Distributor you wish to sell to.
     * (A numbered list of Distributors with their Names, Addresses and Summaries is displayed)
 10. Confirm Transaction. Enter "Yes" or "No"
     * (Message is displayed indicating the status of request.)
