+++ src/main/java/Entities/Farmer.java
@@ -1,7 +1,7 @@
-/* CRC Class Information -------
+package Entities;/* CRC Class Information -------
 
-Class name: Farmer
-Parent Class: User
+Class name: Entities.Farmer
+Parent Class: Entities.User
 
 Responsibilities:
 * Stores farmer’s list of requests
@@ -9,7 +9,7 @@ Responsibilities:
 
 Collaborators:
 * MatchSystem
-* ProfileManager
+* UseCases.ProfileManager
 
 ------------------------------*/
