+++ src/main/java/Entities/User.java
@@ -1,22 +1,22 @@
-/* CRC Class Information -------
+package Entities;/* CRC Class Information -------
 
-Class name: User
+Class name: Entities.User
 Parent Class: None
 
 Responsibilities:
 * Stores name, address, identification, summary
 
 Collaborators:
-* ProfileManager
-* Farmer
-* Distributor
+* UseCases.ProfileManager
+* Entities.Farmer
+* Entities.Distributor
 
 ------------------------------*/
 
 
 public class User {
 
-    //User's Name & Identification
+    //Entities.User's Name & Identification
     //There may be more information added(i.e. contact)
     private String user_name;
     private final int user_id; //(1)
@@ -29,7 +29,7 @@ public class User {
     public String summary;
 
      /**
-     * Construct a User, giving them the name and address
+     * Construct a Entities.User, giving them the name and address
      *
      * @param user_name String for name
      * @param user_address  String for address
