+++ src/main/java/UseCases/MatchManager.java
@@ -1,3 +1,7 @@
+package UseCases;
+
+import Entities.Distributor;
+
 import java.util.ArrayList;
 
 /*
@@ -13,13 +17,13 @@ import java.util.ArrayList;
  * * Create offer using distributor's information [expansion]
  *
  * Collaborators:
- *  Farmer
- *  Distributor
+ *  Entities.Farmer
+ *  Entities.Distributor
  *  RankingSystem
- *  ProfileManager
+ *  UseCases.ProfileManager
  *  Request
  *  Offer
- *  ServiceController
+ *  Controller.ServiceController
  */
 
 public class MatchManager {
