+++ src/test/java/DistributorTest.java
@@ -1,3 +1,4 @@
+import Entities.Distributor;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
