+++ src/test/java/FarmerTest.java
@@ -1,3 +1,4 @@
+import Entities.Farmer;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
