+++ src/test/java/RequestTest.java
@@ -1,3 +1,4 @@
+import Entities.Request;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
