+++ src/test/java/UserTest.java
@@ -1,3 +1,4 @@
+import Entities.User;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
